refactor: rename package
[member -> dto]
